Vulkan: Buffer - Created staging buffer
The host accessiable buffer may not be the best for GPU.
So, created two buffers: one host buffer for CPU to copy data,
the other one is gpu buffer which used for storing the buffer
in GPU optimized format. We need to copy the buffer from host
buffer to GPU buffer if we want to updata buffer data.

- Created staging buffer for using GPU optimized format buffer
  to improve performance
- Created buffer copy command in command manager
- Created staging buffer and copy the buffer to GPU optimized buffer